WIP: Build binary expression tree out of numeral
With this change the numeral to number conversion works by building a binary expression tree out of the numeral. The numeral itself can be seen like an arithmetic problem. As such it consists of additions and multiplications, however, the order is not so straightforward; parenthesis have to be added implicitly and partial calculations be aggregated correctly.