Add support for an implicit leading "one" in numerals
Numerals like "thousand" or "hundred" are now supported without having to explictely lead them with "one".
Additionally, the word "a" as in "a million" is now also supported as the first word in a positive numeral or the second word in a negative numeral (i.e. a numeral that starts with "negative" or "minus").