Allow decimal and thousands separator symbols to change
This makes it possible to change decimal and thousands separator symbols after a converter was instantiated. This comes with the detriment that the first number to numeral conversion will be slower after changing either the decimal or thousands separator symbol, since a new number pattern regular expression has to be compiled.