Optimize numeral pattern regular expression
This makes sure that numerals that consists of only `negative` or `minus` are not treated as valid. Additionally, numerals that contain actual numbers are required to group either digits [0-9] or letters [a-z] together and have these groups be separated by white spaces; as such `8million` is not valid whereas `8 million` is.